Customer NavLink almost works; surrounding area responds to click and animates but only the navlink text actually pushes destination so far
